Created a new widget TextControl
Shifted the logic of button from `main.dart` file to this new file `text_control.dart`. Right now this widget takes in a function pointer which it uses to handle the `onPressed` of the button.

Lateron we can also configure this widget to take in text from the parent widget