Splitted the app into three widgets: App, TextControl & DisplayText
The logic to control the text was shifted from `main.dart` file to `text_control.dart` file,

also the logic to display text was shifted from `main.dart` file to `display_text.dart` file.

As of now this was a very minor change, but this will be helpful later on.